pyproject.toml
[tool]
[tool.poetry]
name = "smunger"
version = "0.0.6"
homepage = "https://github.com/jianhua/smunger"
description = "munger for GWAS summary statistics."
authors = ["Jianhua Wang <jianhua.mert@gmail.com>"]
readme = "README.md"
license =  "MIT"
classifiers=[
    'Development Status :: 2 - Pre-Alpha',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Natural Language :: English',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
]
packages = [
    { include = "smunger" },
    { include = "tests", format = "sdist" },
]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
typer = "^0.7.0"

black = {version = ">=22.3.0", optional = true}
isort  = { version = "^5.8.0", optional = true}
flake8  = { version = "^3.9.2", optional = true}
flake8-docstrings = { version = "^1.6.0", optional = true }
mypy = {version = "^0.900", optional = true}
pytest  = { version = "^6.2.4", optional = true}
pytest-cov  = { version = "^2.12.0", optional = true}
tox  = { version = "^3.20.1", optional = true}
virtualenv  = { version = "^20.2.2", optional = true}
pip  = { version = "^20.3.1", optional = true}
twine  = { version = "^3.3.0", optional = true}
pre-commit = {version = "^2.12.0", optional = true}
toml = {version = "^0.10.2", optional = true}
bump2version = {version = "^1.0.1", optional = true}
mkdocs = {version = "^1.4.2", optional = true}
mkdocs-include-markdown-plugin = {version = "^4.0.3", optional = true}
mkdocs-material = {version = "^8.5.11", optional = true}
mkdocs-autorefs = {version = "^0.4.1", optional = true}
mkdocstrings = {version = "^0.19.1", extras = ["python"], optional = true}
mkdocs-material-extensions = {version = "^1.1.1", optional = true}
jupyter = { version = "^1.0.0", optional = true }
pandas = "^1.5.3"
rich = "^13.3.1"
liftover = "^1.1.16"
requests = "^2.28.2"
biomart = "^0.9.2"
pytabix = "^0.1"

[tool.poetry.extras]
test = [
    "pytest",
    "black",
    "isort",
    "mypy",
    "flake8",
    "flake8-docstrings",
    "pytest-cov"
    ]

dev = ["tox", "pre-commit", "virtualenv", "pip", "twine", "toml", "bump2version"]

doc = [
    "mkdocs",
    "mkdocs-include-markdown-plugin",
    "mkdocs-material",
    "mkdocstrings",
    "mkdocs-material-extension",
    "mkdocs-autorefs"
    ]


[tool.poetry.scripts]
smunger = 'smunger.cli:app'

[tool.black]
line-length = 120
skip-string-normalization = true
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
skip_gitignore = true
# you can skip files as below
#skip_glob = docs/conf.py

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
